requirements.txt
pytest
pytest-html
Appium-Python-Client